r490: - expanded the test suite to test modify and delete operations
- made yet another attempt to make ldb const clean. 

- "make test" now runs both the tdb and ldap backend tests, and run the ldbtest utility
  with and without indexing

- added prototypes in ldb.h for ldb_msg_*() public functions

+++ source/lib/ldb/tests/init_slapd.sh
@@ -5,7 +5,11 @@ export PATH=/usr/sbin:$PATH
 rm -rf tests/tmp/db
 mkdir -p tests/tmp/db
 
-killall slapd
+if pidof slapd > /dev/null; then
+    killall slapd
+fi
 sleep 2
-killall -9 slapd
+if pidof slapd > /dev/null; then
+    killall -9 slapd
+fi
 slapadd -f tests/slapd.conf < tests/init.ldif || exit 1
